Improve validation and code descriptions. Fix some typos.

+++ srcs/coolant_controller.h
@@ -1,8 +1,10 @@
 /***********************************************************************
 * Filename: coolant_controller.h
-* File description: Proportional integral Derivative controller.
-* Used formula is
-* u(t) = kp * temp_error + ki * integral + kd * derivative
+* File description: This file process main logic of cooler and
+* regulates pump and fan power using
+* Proportional integral Derivative controller. 
+* Used formula is:
+* ` u(t) = kp * temp_error + ki * integral + kd * derivative `
 *
 * Project: EAE_Firmware
 * Author: Kateryna Oliinyk
@@ -12,7 +14,7 @@
 #ifndef _PID_CONTROLLER_H_
 #define _PID_CONTROLLER_H_
 
-/* Celcius */
+/* Celsius */
 #define DCDC_TEMPERATURE_LIMIT 110
 
 typedef struct pid_controller_s {
@@ -20,7 +22,7 @@ typedef struct pid_controller_s {
     const int8_t ki;
     const int8_t kd;
     int8_t prev_error;
-    int8_t integral;
+    int16_t integral;
 } pid_controller_t;
 
 /* Wrapper over drives function drv_ignition_switch to achive proper error handling. */
